Migrate check in functionality
